+++ EliteAPI/Features/Badges/DeleteBadge/Endpoint.cs
@@ -11,38 +11,32 @@ internal sealed class DeleteBadgeEndpoint(
 	DataContext context,
 	IObjectStorageService objectStorageService,
 	IOutputCacheStore cacheStore
-	) : Endpoint<BadgeRequest> 
-{
+) : Endpoint<BadgeRequest> {
 	public override void Configure() {
 		Delete("/badge/{BadgeId}");
 		Policies(ApiUserPolicies.Admin);
 		Version(0);
-		
-		Summary(s => {
-			s.Summary = "Delete a badge";
-		});
+
+		Summary(s => { s.Summary = "Delete a badge"; });
 	}
 
-	public override async Task HandleAsync(BadgeRequest request, CancellationToken c) 
-	{
+	public override async Task HandleAsync(BadgeRequest request, CancellationToken c) {
 		var existingBadge = await context.Badges
 			.Include(b => b.Image)
-			.FirstOrDefaultAsync(b => b.Id == request.BadgeId, cancellationToken: c);
-    
+			.FirstOrDefaultAsync(b => b.Id == request.BadgeId, c);
+
 		if (existingBadge is null) {
 			await Send.NotFoundAsync(c);
 			return;
 		}
-		
+
 		context.Badges.Remove(existingBadge);
-		
-		if (existingBadge.Image is not null) {
-			await objectStorageService.DeleteAsync(existingBadge.Image.Path, c);
-		}
-		
+
+		if (existingBadge.Image is not null) await objectStorageService.DeleteAsync(existingBadge.Image.Path, c);
+
 		await context.SaveChangesAsync(c);
 		await cacheStore.EvictByTagAsync("badges", c);
-		
-		await Send.NoContentAsync(cancellation: c);
+
+		await Send.NoContentAsync(c);
 	}
 }
