+++ EliteAPI/Features/Badges/UpdateBadge/Endpoint.cs
@@ -11,50 +11,45 @@ internal sealed class UpdateBadgeEndpoint(
 	DataContext context,
 	IObjectStorageService objectStorageService,
 	IOutputCacheStore cacheStore
-	) : Endpoint<UpdateBadgeRequest> 
-{
+) : Endpoint<UpdateBadgeRequest> {
 	public override void Configure() {
 		Patch("/badge/{BadgeId}");
 		Policies(ApiUserPolicies.Admin);
 		Version(0);
-		
+
 		AllowFileUploads();
 		AllowFormData();
 
-		Summary(s => {
-			s.Summary = "Update a badge";
-		});
+		Summary(s => { s.Summary = "Update a badge"; });
 	}
 
-	public override async Task HandleAsync(UpdateBadgeRequest request, CancellationToken c) 
-	{
+	public override async Task HandleAsync(UpdateBadgeRequest request, CancellationToken c) {
 		var existingBadge = await context.Badges
 			.Include(b => b.Image)
-			.FirstOrDefaultAsync(b => b.Id == request.BadgeId, cancellationToken: c);
-    
+			.FirstOrDefaultAsync(b => b.Id == request.BadgeId, c);
+
 		if (existingBadge is null) {
 			await Send.NotFoundAsync(c);
 			return;
 		}
-    
+
 		existingBadge.Name = request.Badge.Name ?? existingBadge.Name;
 		existingBadge.Description = request.Badge.Description ?? existingBadge.Description;
 		existingBadge.Requirements = request.Badge.Requirements ?? existingBadge.Requirements;
-    
+
 		if (request.Badge.Image is not null) {
-			if (existingBadge.Image is not null) {
-				await objectStorageService.DeleteAsync(existingBadge.Image.Path, c);
-			}
-        
-			var image = await objectStorageService.UploadImageAsync($"badges/{existingBadge.Id}/{Guid.NewGuid()}.png", request.Badge.Image, token: c);
-       
+			if (existingBadge.Image is not null) await objectStorageService.DeleteAsync(existingBadge.Image.Path, c);
+
+			var image = await objectStorageService.UploadImageAsync($"badges/{existingBadge.Id}/{Guid.NewGuid()}.png",
+				request.Badge.Image, token: c);
+
 			existingBadge.Image = image;
 			existingBadge.ImageId = image.Id;
 		}
-    
+
 		await context.SaveChangesAsync(c);
 		await cacheStore.EvictByTagAsync("badges", c);
-		
-		await Send.NoContentAsync(cancellation: c);
+
+		await Send.NoContentAsync(c);
 	}
 }
